requirements.txt
bcrypt==4.1.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1